AGS: Fixed MemoryStream::ReadByte()
From upstream 07c7f8492771cffa9b6831c0a56b2e8a7fdd5410

+++ engines/ags/shared/util/memory_stream.cpp
@@ -110,7 +110,7 @@ int32_t MemoryStream::ReadByte() {
 	if (EOS()) {
 		return -1;
 	}
-	return _cbuf[(size_t)(_pos++)];
+	return static_cast<uint8_t>(_cbuf[(size_t)(_pos++)]);
 }
 
 size_t MemoryStream::Write(const void *buffer, size_t size) {
